Exercise 4.21

+++ BlogList/controllers/blogs.js
@@ -13,34 +13,32 @@ blogRouter.get('/api/blogs', async (request,response) => {
 
 blogRouter.post('/api/blogs', async (request, response) => {
 
-    try {
-        const body = request.body 
 
-        const decodedToken = jwt.verify(request.token, process.env.SECRET)
-        if (!decodedToken || !request.token) {
-            return response.status(401).json({ error: 'token invalid' }) 
-        }
-        const user = await User.findById(decodedToken.id)    
+    const body = request.body 
 
-        if (!body.title || !body.url){
-            return response.status(400).json({error: "Title or URL missing"})
-        }
-
-        const blog = new Blog({
-                title: body.title,
-                author: body.author,
-                url: body.url,
-                likes: body.likes ? body.likes : 0
-        })
-
-        blog.user = user.id
-        const savedBlog = await blog.save()
-        user.blogs = user.blogs.concat(savedBlog.id)
-        await user.save()
-        response.status(201).json(savedBlog)
-    } catch (error){
-        errorHandler(error)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken) {
+        return response.status(401).json({ error: 'token invalid' }) 
     }
+    const user = await User.findById(decodedToken.id)    
+
+    if (!body.title || !body.url){
+        return response.status(400).json({error: "Title or URL missing"})
+    }
+
+    const blog = new Blog({
+            title: body.title,
+            author: body.author,
+            url: body.url,
+            likes: body.likes ? body.likes : 0
+    })
+
+    blog.user = user.id
+    const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog.id)
+    await user.save()
+    response.status(201).json(savedBlog)
+
     
 })
 
@@ -58,12 +56,23 @@ blogRouter.get('/api/blogs/:id', async (request,response) => {
 blogRouter.delete('/api/blogs/:id', async (request,response) => {
     const id = request.params.id
 
-    try {
-        const blog = await Blog.findByIdAndDelete(id)
-        response.json(blog)
-    } catch (error) {
-        response.status(400).json({error})
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken) {
+        return response.status(401).json({ error: 'token invalid' }) 
     }
+
+    const userTokenId = decodedToken.id
+
+    const blog = await Blog.findById(id)
+    
+    if (userTokenId === blog.user.toString()){
+        blogToDelete = await Blog.findByIdAndDelete(id)
+        response.json(blogToDelete)
+    } else {
+        response.status(400).json({ error: 'Invalid user trying to delete blog' }) 
+    }
+    response.status(200).end()
+
 })
 
 blogRouter.put('/api/blogs/:id', async (request,response) => {
